+++ src/pages/author-detail/components/AuthorArtwork/index.style.ts
