+++ src/pages/author-detail/components/AuthorExhibition/index.style.ts
