+++ src/pages/author-detail/components/AuthorExhibition/index.tsx
